+++ CyberPets.Domain/UserPets/UserPetEntity.cs
@@ -29,10 +29,10 @@ namespace CyberPets.Domain.UserPets
 
         public int GetHunger(DateTime now) => HungerMetric.GetValue(now, Kind.HungerRateInSeconds);
 
-        public UserPetMetricValue UpdatedHunger(DateTime now, double amount) => HungerMetric.Updated(now, Kind.HungerRateInSeconds, -amount);
+        public UserPetMetricValue UpdatedHunger(DateTime now, int amount) => HungerMetric.Updated(now, Kind.HungerRateInSeconds, -amount);
 
         public int GetHappiness(DateTime now) => HappinessMetric.GetValue(now, -Kind.HappinessRateInSeconds);
 
-        public UserPetMetricValue UpdatedHappiness(DateTime now, double amount) => HappinessMetric.Updated(now, -Kind.HappinessRateInSeconds, amount);
+        public UserPetMetricValue UpdatedHappiness(DateTime now, int amount) => HappinessMetric.Updated(now, -Kind.HappinessRateInSeconds, amount);
     }
 }
